chore: moved environment selection to EntryPoint
