+++ Explorer/Assets/Scripts/Global/Dynamic/DynamicSceneLoaderSettings.cs
@@ -9,20 +9,8 @@ namespace Global.Dynamic
     [CreateAssetMenu(fileName = "DynamicSceneLoaderSettings", menuName = "DCL/Various/Dynamic Scene LoaderSettings")]
     public class DynamicSceneLoaderSettings : ScriptableObject
     {
-        [field: SerializeField] public DecentralandEnvironment DecentralandEnvironment { get; private set; }
         [field: SerializeField] public List<string> Realms { get; private set; }
         [field: SerializeField] public List<string> Web3WhitelistMethods { get; private set; }
 
-        public void ApplyConfig(IAppArgs applicationParametersParser)
-        {
-            if (applicationParametersParser.TryGetValue(AppArgsFlags.ENVIRONMENT, out string? environment))
-                ParseEnvironment(environment!);
-        }
-
-        private void ParseEnvironment(string environment)
-        {
-            if (Enum.TryParse(environment, true, out DecentralandEnvironment env))
-                DecentralandEnvironment = env;
-        }
     }
 }
